Initial commit for Samsung Galaxy Tab 3 7.0" Sprint variant (lt02spr) and possibly others.
Some notable details:
This device is an LTE device and uses the same radio technology (WCNSS) as the apextmo, expressatt, goghspr
and infinitussc, which are msm8960l devices. YAS sensors, also in the listed devices, are present. Optical sensor
matches that from the d2dcm. There are several other common board devices present.